http: never read more from a response than specified in content-length
We don't really need this for now, because we explicitly tell the server
that we don't support persistent connections and on top of that we close
the write end of the socket right after sending a request. However, we
might want to change this one day and with this patch we will be ready.
Besides, knowing content length will be useful for drawing progress bar,
and it's a good cause to introduce infrastructure for handling headers,
which we will need later for supporting chunked responses.

Signed-off-by: Vladimir Davydov <vdavydov@me.com>